+++ api/attraction_api.py
@@ -1,22 +1,22 @@
 from flask import Blueprint ,  request
-import mysql_function 
+from api.models.attraction import Attraction
 
 
 attraction_api = Blueprint("attraction_api", __name__)
-getData = mysql_function.ConnectToSql()
+attractions = Attraction()
 
 @attraction_api.route("/api/attraction/<id>")
 def attraction(id):
 	id = int(id)
 	try:
-		if(len(getData.get_attraction(id)) == 0):
+		if(len(attractions.get_attraction(id)) == 0):
 			data = {
 				"error":True,
 				"message":"景點編號不正確"
 			}
 			return data,400
 		data = {
-			"data":getData.get_attraction(id)
+			"data":attractions.get_attraction(id)
 		}
 		return data,200
 	except:
@@ -34,7 +34,7 @@ def apiAttraction():
 		page = request.args.get("page","0")
 		keyword = request.args.get("keyword", "")
 		page=int(page)
-		if(len(getData.show_page(page+1)) == 0):
+		if(len(attractions.show_page(page+1)) == 0):
 			nextPage = None
 		else:
 			nextPage = page +1
@@ -46,18 +46,18 @@ def apiAttraction():
 			}
 			return data,200
 		if(keyword != ""):
-			if(len(getData.show_page(page+1,keyword)) == 0):
+			if(len(attractions.show_page(page+1,keyword)) == 0):
 				nextPage = None
 				
 			data = {
 				"nextPage":nextPage,
-				"data":getData.show_page(page,keyword)
+				"data":attractions.show_page(page,keyword)
 			}
 			return data,200
 		if(keyword == ""):
 			data = {
 					"nextPage":nextPage,
-					"data":getData.show_page(page)
+					"data":attractions.show_page(page)
 			}
 			return data,200
 	except:
