+++ static/styles/booking.css
@@ -216,7 +216,7 @@ input{
 .empty-shoppingcart{
     display:none;
     margin:auto;
-    margin-top:50px;
+    margin-top:70px;
 }
 
 @media(max-width:1200px){
